feat: add generate theme script (#1219)
* feat: add generate theme script

* refactor: structure
