+++ packages/static/src/breadcrumbs/styles/breadscrumbs.module.css
@@ -1,4 +1,4 @@
-@import '../../theme-static/theme.module.css';
+@import '../../theme/theme.module.css';
 
 .breadcrumbs {
   font-family: var(--cdg-font-sans);
