+++ packages/static/tsconfig.json
@@ -21,5 +21,5 @@
     "noFallthroughCasesInSwitch": false
   },
   "include": ["src"],
-  "references": [{ "path": "./tsconfig.node.json" }]
+  "references": [{"path": "./tsconfig.node.json"}]
 }
